:sparkles: refactor : exceptions 디렉토리 이동
