+++ src/main/kotlin/org/ekgns33/springmarket/auth/exceptions/UserAuthenticationException.kt
@@ -1,6 +1,6 @@
 package org.ekgns33.springmarket.auth.exceptions
 
-import org.ekgns33.springmarket.common.GlobalErrorCode
+import org.ekgns33.springmarket.common.exceptions.GlobalErrorCode
 
 class UserAuthenticationException(
     val errorCode: GlobalErrorCode,
