--- src/main/kotlin/org/ekgns33/springmarket/common/BusinessServiceException.kt
@@ -1,5 +1,0 @@
-package org.ekgns33.springmarket.common
-
-open class BusinessServiceException(
-    val errorCode: ErrorCode) : RuntimeException() {
-}
