+++ src/main/kotlin/org/ekgns33/springmarket/common/exceptions/BusinessServiceException.kt
@@ -1,0 +1,6 @@
+package org.ekgns33.springmarket.common.exceptions
+
+open class BusinessServiceException(
+    val errorCode: ErrorCode
+) : RuntimeException() {
+}
