+++ src/main/kotlin/org/ekgns33/springmarket/common/exceptions/ErrorCode.kt
@@ -1,4 +1,4 @@
-package org.ekgns33.springmarket.common
+package org.ekgns33.springmarket.common.exceptions
 
 import org.springframework.http.HttpStatusCode
