+++ src/main/kotlin/org/ekgns33/springmarket/product/exceptions/ProductErrorCode.kt
@@ -1,6 +1,6 @@
 package org.ekgns33.springmarket.product.exceptions
 
-import org.ekgns33.springmarket.common.ErrorCode
+import org.ekgns33.springmarket.common.exceptions.ErrorCode
 import org.springframework.http.HttpStatus
 
 enum class ProductErrorCode(
