+++ src/main/kotlin/org/ekgns33/springmarket/product/exceptions/ProductNotFoundException.kt
@@ -1,6 +1,6 @@
 package org.ekgns33.springmarket.product.exceptions
 
-import org.ekgns33.springmarket.common.BusinessServiceException
-import org.ekgns33.springmarket.common.ErrorCode
+import org.ekgns33.springmarket.common.exceptions.BusinessServiceException
+import org.ekgns33.springmarket.common.exceptions.ErrorCode
 
 class ProductNotFoundException(errorCode: ErrorCode) : BusinessServiceException(errorCode)
